+++ monitor/amqp/client.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 """
-Rabbit MQ Client
-================
+RabbitMQ AMQP Client
+====================
 Modified: 2021-10
 """
+
 import json
+import logging
 from typing import Any, Dict
 from pika import BlockingConnection, ConnectionParameters, PlainCredentials
-import logging
+from monitor.sys import system
 from monitor.events.registry import Registry as events
 from monitor.environment.thread_manager import ThreadManager as tm
 from monitor.logs.formatter import pformat
@@ -17,21 +19,17 @@ from monitor.amqp.conf import AMQPConf
 class AMQPClient:
     def __init__(self, host: str, port: int, username: str, password: str) -> None:
         self._logger = logging.getLogger(__name__)
+        system.wfi(host, port, timeout=30)
         events.amqp_publish.register(
             callback=self.publish,
             priority=1
-        )
-        credentials = PlainCredentials(
-            username="microservice",
-            password="microservice",
-            erase_on_connect=True
         )
         self.connection = BlockingConnection(
             ConnectionParameters(
                 host=host,
                 port=port,
                 virtual_host='/',
-                credentials=credentials
+                credentials=PlainCredentials(username, password, erase_on_connect=True)
             )
         )
         self._logger.info("Instantiation successful.")
