+++ monitor/environment/state_manager.py
@@ -34,7 +34,7 @@ from monitor.models.experiment import Experiment
 from monitor.models.imaging_profile import ImagingProfile
 from monitor.environment.registry import CallbackRegistry as cr
 from monitor.environment.registry import StateRegistry as sr
-from monitor.sys.helpers import clear_thumbnail, write_lab_id, read_lab_id
+from monitor.sys.system import clear_thumbnail, write_lab_id, read_lab_id
 
 StateModel = Union[ICB, Experiment, Device, Protocol, ImagingProfile]
 
@@ -184,7 +184,7 @@ class StateManager:
         :param icb: icb runtime model
         :type icb: ICB
         """
-        sr.icb.setattrs(**icb.__dict__)
+        sr.icb.serialize(**icb.__dict__)
 
     def _load_runtime_models(self) -> None:
         """
